Rename iRegistrationRequestBody file
